+++ flows/base/create_cachedb_file_plugin/copy.py
@@ -1,15 +1,32 @@
-import traceback
+import duckdb
 from typing import Any
+from psycopg2 import connect
 
 from prefect import task
+from prefect.variables import Variable
+from prefect.blocks.system import Secret
+from prefect.context import TaskRunContext
 from prefect.logging import get_run_logger
+from prefect.tasks import exponential_backoff
 
-from .fts import create_fts_index
 from .types import CopyParameters, QueryColumns, _COPY_STATUS_TABLE_NAME
-from .utils import execute_statement, set_bigquery_global_settings, VOCAB_TABLES
 from .filter import filter_tables, _CDM_COLUMN_FILTER_MAP, _CHUNK_COLUMN_MAP
+from .utils import execute_statement, set_bigquery_global_settings, VOCAB_TABLES
+from .chunk_utils import determine_chunk_size, plan_chunks, find_column_case_insensitive
+
 from _shared_flow_utils.types import SupportedDatabaseDialects
-import sqlalchemy as sql
+
+
+def get_trex_connection(database_code: str):    
+    conn = connect(
+        host=Variable.get("trex_sql_host"),
+        port=Variable.get("trex_sql_port"),
+        user=Variable.get("trex_sql_user"),
+        password=Secret.load("trex-sql-password").get(),
+        dbname=database_code,
+    )
+    conn.autocommit = True
+    return conn
 
 
 @task(log_prints=True, task_run_name="create_cache_status_table")
@@ -43,6 +60,7 @@ def mark_complete(con, table: str, copy_params):
         """
     )
 
+
 def cleanup(con, table: str, copy_params):
     execute_statement(con, f"DROP TABLE IF EXISTS \"{copy_params.target_database}\".\"{copy_params.target_schema}\".\"{table}\"")
     execute_statement(con, f"""
@@ -52,56 +70,90 @@ def cleanup(con, table: str, copy_params):
         """
     )
 
+
 @task(log_prints=True, task_run_name="drop_cache_status_table")
 def drop_cache_status_table(con, copy_params):
     execute_statement(con, f'DROP TABLE "{copy_params.target_database}"."{copy_params.target_schema}"."{_COPY_STATUS_TABLE_NAME}";') 
 
 
-@task(log_prints=True, task_run_name="copy_all_schemas_from_{read_conn.database_code}")
-def copy_all_schemas(write_conn: Any, read_conn: Any, copy_params: CopyParameters):
+
+@task(retries=3, 
+      retry_delay_seconds=exponential_backoff(backoff_factor=2),
+      log_prints=True, 
+      task_run_name="create_schema_if_not_exists_{copy_params.target_schema}")
+def create_schema_if_not_exists_task(use_trex_conn: bool, copy_params: CopyParameters, duckdb_file_path: str):
     logger = get_run_logger()
-    logger.info(f"Starting schema copy for database '{read_conn.database_code}'...")
-    schemas_to_copy = sorted(read_conn.get_schema_names())
-    logger.info(f"Found {len(schemas_to_copy)} schemas: {schemas_to_copy}")
 
-    failed_schemas = []
+    task_run_ctx = TaskRunContext.get()
+    logger.info(f"This is task run attempt: {task_run_ctx.task_run.run_count} for task '{task_run_ctx.task.name}'.")
 
-    for idx, schema in enumerate(schemas_to_copy, start=1):
-        logger.info(f"[{idx}/{len(schemas_to_copy)}] Copying schema '{schema}'...")
+    if use_trex_conn:
+        trex_conn = None
+        pg_cursor = None
+    
         try:
-            create_schema_if_not_exists(write_conn, copy_params)
-            create_schema_tables(write_conn, read_conn, copy_params)
+            trex_conn = get_trex_connection(copy_params.target_database)
+            pg_cursor = trex_conn.cursor()
+            pg_cursor.execute("CALL pg_clear_cache();")
+
+            create_schema_if_not_exists(pg_cursor, copy_params, logger)
+
         except Exception as e:
-            logger.error(f"Failed to copy schema '{schema}': {e}")
-            logger.error(traceback.format_exc())
-            failed_schemas.append(schema)
-            continue
+            logger.error(f"Failed to create schema through Trex SQL interface: {e}")
+            raise
+        finally:
+            if pg_cursor:
+                pg_cursor.close()
+            if trex_conn:
+                trex_conn.close()
 
-        try:
-            create_fts_index(write_conn, read_conn, copy_params)
-        except Exception as e:
-            logger.error(f"Failed to create FTS index for schema '{schema}': {e}")
-            logger.error(traceback.format_exc())
-            failed_schemas.append(schema)
+    else:
+        with duckdb.connect(duckdb_file_path) as file_conn:
+            create_schema_if_not_exists(file_conn, copy_params, logger)
 
-    total = len(schemas_to_copy)
-    success = total - len(failed_schemas)
-    logger.info(f"Finished copying schemas: Total={total}, Successful={success}, Failed={len(failed_schemas)}")
-    if failed_schemas:
-        logger.error(f"Schemas failed: {', '.join(failed_schemas)}")
 
-@task(log_prints=True, task_run_name="create_schema_if_not_exists_{copy_params.target_schema}")
-def create_schema_if_not_exists(write_conn: Any, copy_params: CopyParameters):
-    logger = get_run_logger()
+def create_schema_if_not_exists(write_conn: Any, copy_params: CopyParameters, logger):
+    logger.info(f"Starting creation of schema '{copy_params.target_schema}' in database '{copy_params.target_database}' if it doesn't exist...")
     sql = f'CREATE SCHEMA IF NOT EXISTS "{copy_params.target_database}"."{copy_params.target_schema}";'
     execute_statement(write_conn, sql)
-    logger.info(f"Schema '{copy_params.target_schema}' ensured.")
+    logger.info(f"Schema '{copy_params.target_schema}' created.")
 
-@task(log_prints=True, task_run_name="create_schema_tables_from_{copy_params.source_schema}")
-def create_schema_tables(write_conn: Any, read_conn: Any, copy_params: CopyParameters):
+
+@task(retries=3, 
+      retry_delay_seconds=exponential_backoff(backoff_factor=2),
+      log_prints=True, 
+      task_run_name="create_schema_tables_from_{copy_params.source_schema}")
+def create_schema_tables_task(use_trex_conn: bool, read_conn: Any, copy_params: CopyParameters, duckdb_file_path: str):
     logger = get_run_logger()
-    logger.info(f"Starting creation of schema '{copy_params.target_schema}' in database '{copy_params.target_database}' if it doesn't exist...")
 
+    task_run_ctx = TaskRunContext.get()
+    logger.info(f"This is task run attempt: {task_run_ctx.task_run.run_count} for task '{task_run_ctx.task.name}'.")
+
+    if use_trex_conn:
+        trex_conn = None
+        pg_cursor = None
+    
+        try:
+            trex_conn = get_trex_connection(copy_params.target_database)
+            pg_cursor = trex_conn.cursor()
+
+            create_schema_tables(pg_cursor, read_conn, copy_params, logger)
+
+        except Exception as e:
+            logger.error(f"Failed to copy schema tables through Trex SQL interface: {e}")
+            raise
+        finally:
+            if pg_cursor:
+                pg_cursor.close()
+            if trex_conn:
+                trex_conn.close()
+
+    else:
+        with duckdb.connect(duckdb_file_path) as file_conn:
+            create_schema_tables(file_conn, read_conn, copy_params, logger)
+
+
+def create_schema_tables(write_conn: Any, read_conn: Any, copy_params: CopyParameters, logger):
     source_schema = copy_params.source_schema
 
     # Create status table
@@ -161,11 +213,11 @@ def create_schema_tables(write_conn: Any, read_conn: Any, copy_params: CopyParam
 
     # Filter out already completed tables
     original_count = len(tables_to_copy)
-    tables_to_copy = [t for t in tables_to_copy if t not in completed_tables]
-    skipped_count = original_count - len(tables_to_copy)
-    logger.info(f"Found {len(tables_to_copy)} tables to copy in schema '{source_schema}'.")
-    logger.info(f"Total tables in source schema: {original_count}")
-    logger.info(f"Tables to copy: {tables_to_copy}")
+    tables_left_to_copy = [t for t in tables_to_copy if t not in completed_tables]
+    skipped_count = original_count - len(tables_left_to_copy)
+    logger.info(f"Found {len(tables_left_to_copy)} tables to copy from schema(s): {copy_params.source_schema}{', ' + copy_params.vocab_schema if has_separate_vocab_schema else ''} .")
+    logger.info(f"Total tables in source schemas: {original_count}")
+    logger.info(f"Tables left to copy: {tables_left_to_copy}")
     if skipped_count > 0:
         logger.info(f"Skipping {skipped_count} already completed tables.")
 
@@ -209,6 +261,7 @@ def create_schema_tables(write_conn: Any, read_conn: Any, copy_params: CopyParam
     # All tables copied successfully, drop the status tracking table
     drop_cache_status_table(write_conn, copy_params)
 
+
 def create_empty_target_table(write_conn: Any, copy_params: CopyParameters, query_columns: QueryColumns, source_schema: str):
     select_sql = create_select_query(copy_params, query_columns, source_schema, None)
     execute_statement(write_conn, f'DROP TABLE IF EXISTS "{copy_params.target_database}"."{copy_params.target_schema}"."{query_columns.table}";')
@@ -218,68 +271,6 @@ def create_empty_target_table(write_conn: Any, copy_params: CopyParameters, quer
     """
     execute_statement(write_conn, sql)
 
-def determine_chunk_size(dialect: str, row_count: int | None, chunk_size: int | None = None) -> int:
-    if chunk_size is not None:
-        return chunk_size
-    if dialect == SupportedDatabaseDialects.BIGQUERY.value:
-        return 5_000_000
-    if row_count and row_count > 100_000_000:
-        return 1_000_000
-    return 1_000_000
-
-def plan_chunks(read_conn: Any, database: str, schema: str, table: str, chunk_col: str, chunk_size: int, row_count: int | None, logger=None):
-    min_val = None
-    max_val = None
-
-    try:
-        # Determine table path and column quoting based on dialect
-        dialect = read_conn.tenant_configs.dialect
-        if dialect == SupportedDatabaseDialects.BIGQUERY.value:
-            table_path = f'`{schema}.{table}`'
-            col_quote = ''
-        else:
-            table_path = f'"{schema}"."{table}"'
-            col_quote = '"'
-        with read_conn.engine.connect() as connection:
-            query = sql.text(f'SELECT MIN({col_quote}{chunk_col}{col_quote}), MAX({col_quote}{chunk_col}{col_quote}) FROM {table_path}')
-            result = connection.execute(query).fetchone()
-            if result:
-                min_val, max_val = result
-        
-    except Exception as e:
-        logger.warning(f"Failed to get min/max for '{table}' from source: {e}")
-        min_val = None
-
-    if min_val is None or max_val is None:
-        logger.info(f"Chunk column '{chunk_col}' has no valid values. Cannot chunk.")
-        return None
-
-    # Try to convert to int; if fails (e.g., for dates), don't chunk
-    try:
-        min_val = int(min_val)
-        max_val = int(max_val)
-    except (ValueError, TypeError):
-        logger.info(f"Chunk column '{chunk_col}' is not numeric. Cannot chunk.")
-        return None
-
-    chunks = []
-    current = min_val
-    while current <= max_val:
-        end = min(current + chunk_size - 1, max_val)
-        chunks.append(f'"{chunk_col}" BETWEEN {current} AND {end}')
-        current = end + 1
-    
-    if row_count and len(chunks) > 0:
-        avg_rows_per_chunk = row_count / len(chunks)
-        if avg_rows_per_chunk > (chunk_size * 2):
-            logger.info(f"Data is too dense for range chunking (avg {int(avg_rows_per_chunk)} rows/chunk vs target {chunk_size}). Cannot chunk efficiently.")
-            return None
-        else:
-            logger.info(f"Planned {len(chunks)} chunks for table '{table}': chunk_size={chunk_size}")
-            return chunks
-    else:
-         logger.info(f"Planned {len(chunks)} chunks for table '{table}': chunk_size={chunk_size}")
-         return chunks
 
 def copy_table_chunk(write_conn: Any, copy_params: CopyParameters, query_columns: QueryColumns, source_schema: str, where_sql: str | tuple, chunk_id: int, total_chunks: int, logger=None):
     logger.info(f"Copying chunk {chunk_id + 1}/{total_chunks} for table '{query_columns.table}'")
@@ -504,10 +495,3 @@ def copy_indexes(write_conn: Any, read_conn: Any, copy_params: CopyParameters, q
                 f"Primary Key Index '{pk_index_name}' copied for table '{table}' in schema '{target_database}'.'{target_schema}'."
             )
 
-def find_column_case_insensitive(columns: list[str], target: str) -> str | None:
-    if not target:
-        return None
-    for col in columns:
-        if col.lower() == target.lower():
-            return col
-    return None
